manage-cluster with support for multiple versions of kubespray
