+++ docker/kubespray_patches/v2.10.0.patch
@@ -1,0 +1,19 @@
+diff --git a/requirements.txt b/requirements.txt
+index eb00af2d..7f71b2e2 100644
+--- a/requirements.txt
++++ b/requirements.txt
+@@ -1,7 +1,7 @@
+-ansible>=2.7.8
+-jinja2>=2.9.6
+-netaddr
+-pbr>=1.6
+-hvac
+-jmespath
+-ruamel.yaml
++ansible==2.7.8
++jinja2==2.10.1
++netaddr==0.7.19
++pbr==5.2.0
++hvac==0.8.2
++jmespath==0.9.4
++ruamel.yaml==0.15.96
