+++ docker/kubespray_patches/v2.8.5.patch
@@ -1,0 +1,1 @@
+v2.8.4.patch
