+++ docker/terraform_openstack_templates/.gitignore
@@ -1,0 +1,4 @@
+.terraform
+*.tfvars
+*.tfstate
+*.tfstate.backup
