+++ docker/terraform_openstack_templates/modules/ips/variables.tf
@@ -1,0 +1,17 @@
+variable "number_of_k8s_masters" {}
+
+variable "number_of_k8s_masters_no_etcd" {}
+
+variable "number_of_k8s_nodes" {}
+
+variable "floatingip_pool" {}
+
+variable "number_of_bastions" {}
+
+variable "external_net" {}
+
+variable "network_name" {}
+
+variable "router_id" {
+  default = ""
+}
