+++ docker/terraform_openstack_templates/modules/network/variables.tf
@@ -1,0 +1,13 @@
+variable "external_net" {}
+
+variable "network_name" {}
+
+variable "cluster_name" {}
+
+variable "dns_nameservers" {
+  type = "list"
+}
+
+variable "subnet_cidr" {}
+
+variable "use_neutron" {}
