+++ docker/terraform_openstack_templates/sample-inventory/group_vars/all/coreos.yml
@@ -1,0 +1,2 @@
+## Does coreos need auto upgrade, default is true
+#coreos_auto_upgrade: true
