+++ docker/terraform_openstack_templates/sample-inventory/group_vars/k8s-cluster/k8s-net-cilium.yml
@@ -1,0 +1,1 @@
+# see roles/network_plugin/cilium/defaults/main.yml
